requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgi-redis==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
Django==3.1.1
django-cors-headers==3.5.0
django-countries==6.1.3
django-group-by==0.3.1
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drf-jwt-2fa==0.3.0
drf-yasg==1.20.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
image==1.5.33
importlib-metadata==3.7.3
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.2
msgpack-python==0.5.6
ndg-httpsclient==0.5.1
packaging==20.4
phonenumbers==8.12.8
Pillow==8.0.1
prompt-toolkit==3.0.18
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-dotenv==0.14.0
pytz==2020.1
qrcode==6.1
redis==3.2.0
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
zope.interface==5.2.0
